Revert "[terra-functional-testing]  Set test page to the viewport size instead of the browser's window size (#673)"
This reverts commit 4ada1f52a5f6315c2256239a696810a717e4a2b0.
