+++ packages/terra-functional-testing/src/commands/utils/index.js
@@ -5,11 +5,9 @@ const dispatchCustomEvent = require('./dispatchCustomEvent');
 const downloadScreenshots = require('./downloadScreenshots');
 const eventEmitter = require('./eventEmitter');
 const getViewports = require('./getViewports');
-const getViewportSize = require('./getViewportSize');
 const hideInputCaret = require('./hideInputCaret');
 const setApplicationLocale = require('./setApplicationLocale');
 const setViewport = require('./setViewport');
-const setViewportSize = require('./setViewportSize');
 
 module.exports = {
   cleanScreenshots,
@@ -19,9 +17,7 @@ module.exports = {
   downloadScreenshots,
   eventEmitter,
   getViewports,
-  getViewportSize,
   hideInputCaret,
   setApplicationLocale,
   setViewport,
-  setViewportSize,
 };
