+++ packages/terra-functional-testing/src/services/wdio-visual-regression-service/modules/getTerraFormFactor.js
@@ -1,5 +1,26 @@
+/* global browser */
+
 const { TERRA_VIEWPORTS } = require('../../../constants');
-const { getViewportSize } = require('../../../commands/utils');
+
+/**
+ * Determines the current viewport size.
+ *
+ * @returns {Object} - the current viewport size.
+ */
+function getViewportSize() {
+  // eslint-disable-next-line prefer-arrow-callback
+  const resolution = browser.execute(function getResolution() {
+    return {
+      screenWidth: Math.max(document.documentElement.clientWidth, window.innerWidth || 0),
+      screenHeight: Math.max(document.documentElement.clientHeight, window.innerHeight || 0),
+    };
+  });
+
+  return {
+    width: resolution.screenWidth,
+    height: resolution.screenHeight,
+  };
+}
 
 /**
  * Determines the Terra form factor for the current viewport size.
